src/requirements.txt
altair==5.0.1
asttokens==2.2.1
attrs==23.1.0
backcall==0.2.0
bcrypt==4.0.1
bleach==6.0.0
blinker==1.6.2
bokeh==3.2.1
cachecontrol==0.13.1
cachetools==5.3.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.6
colorama==0.4.6
contourpy==1.1.0
cryptography==41.0.2
cycler==0.11.0
decorator==5.1.1
et-xmlfile==1.1.0
executing==1.2.0
extra-streamlit-components==0.1.56
firebase-admin==6.2.0
fonttools==4.41.0
gitdb==4.0.10
GitPython==3.1.32
google-api-core==2.11.1
google-api-python-client==2.94.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
google-auth==2.22.0
google-cloud-core==2.3.3
google-cloud-firestore==2.11.1
google-cloud-storage==2.10.0
google-crc32c==1.5.0
google-resumable-media==2.5.0
googleapis-common-protos==1.59.1
grpcio-status==1.56.2
grpcio==1.56.2
httplib2==0.22.0
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.0.0
ipython==8.14.0
jedi==0.18.2
Jinja2==3.1.2
jsonschema-specifications==2023.7.1
jsonschema==4.18.4
kiwisolver==1.4.4
linkify-it-py==2.0.2
markdown-it-py==3.0.0
Markdown==3.4.3
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
matplotlib==3.7.2
mdit-py-plugins==0.4.0
mdurl==0.1.2
msgpack==1.0.5
numpy==1.25.1
oauthlib==3.2.2
openpyxl==3.1.2
packaging==23.1
pandas==2.0.3
panel==1.2.0
param==1.13.0
parso==0.8.3
pickleshare==0.7.5
Pillow==9.5.0
pip==22.3.1
platformdirs==3.9.1
pooch==1.7.0
prompt-toolkit==3.0.39
proto-plus==1.22.3
protobuf==4.23.4
pure-eval==0.2.2
pyarrow==12.0.1
pyasn1-modules==0.3.0
pyasn1==0.5.0
pycparser==2.21
pydeck==0.8.1b0
Pygments==2.15.1
PyJWT==2.8.0
Pympler==1.0.1
pyparsing==3.0.9
python-dateutil==2.8.2
pytz-deprecation-shim==0.1.0.post0
pytz==2023.3
pyvista==0.41.0
pyviz-comms==2.3.2
PyYAML==6.0.1
referencing==0.30.0
requests-oauthlib==1.3.1
requests==2.31.0
rich==13.4.2
rpds-py==0.9.2
rsa==4.9
scooby==0.7.2
setuptools==65.5.1
six==1.16.0
smmap==5.0.0
stack-data==0.6.2
streamlit-authenticator==0.2.2
streamlit==1.25.0
tenacity==8.2.2
toml==0.10.2
toolz==0.12.0
tornado==6.3.2
tqdm==4.65.0
traitlets==5.9.0
typing-extensions==4.7.1
tzdata==2023.3
tzlocal==4.3.1
uc-micro-py==1.0.2
uritemplate==4.1.1
urllib3==1.26.16
uuid==1.30
validators==0.20.0
vtk==9.2.6
watchdog==3.0.0
wcwidth==0.2.6
webencodings==0.5.1
wheel==0.38.4
xyzservices==2023.7.0
zipp==3.16.2
